Removed a lot of warnings from KeyMagicInputMethod and MainFile. In a lot of these cases, warnings implied errors, so I'm glad I took them out. (Also prevents them from hiding future warnings in a huge wall of text).

+++ win32_source/Input/KeyMagicInputMethod.h
@@ -159,7 +159,7 @@ public:
 		return strDot;
 	}
 	bool isDone() {
-		return (rulestream==NULL) || (dot==rulestream->size());
+		return (rulestream==NULL) || (dot==(int)rulestream->size());
 	}
 	const std::vector<Rule>* const getRulestream() {
 		return rulestream;
